requirements.txt
cassandra-driver
psutil
argparse
jmxquery
pyyaml
tqdm